pyproject.toml
[project]
name = "kiwi-mcp"
version = "0.1.0"
description = "Unified MCP for directives, scripts, and knowledge"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["mcp", "ai", "agents", "directives", "scripts", "knowledge"]

dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "aiofiles>=23.0.0",
    "packaging>=24.0",
    "jsonschema>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]
rag = [
    "numpy>=1.20.0",      # For cosine similarity calculations (optional but recommended)
]

[project.scripts]
kiwi-mcp = "kiwi_mcp.server:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["kiwi_mcp*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "version: marks tests related to version checking and warnings",
]

[tool.ruff]
line-length = 100
target-version = "py311"
